Fix showing diagnostic tooltip
Ranges are semiopen, better understood as delimited by caret positions.
SCI_INDICATORVALUEAT will find the indicator for the positions starting
in in range.from up to but not including range.to, which is particularly
important when the range spans just one character.

Also, although it shouldn't be an issue with the range check change,
don't return an index when a diagnostic is not found.

Fixes sometimes displaying an empty diagnostic tooltip and sometimes not
displaying any at all.